+++ database/migrate/migrations/00004_chunk.sql
@@ -19,15 +19,15 @@ create table chunk
     proof                             BYTEA           DEFAULT NULL,
     prover_assigned_at                TIMESTAMP(0)    DEFAULT NULL,
     proved_at                         TIMESTAMP(0)    DEFAULT NULL,
-    proof_time_sec                    SMALLINT        DEFAULT NULL,
+    proof_time_sec                    INTEGER         DEFAULT NULL,
 
 -- batch
     batch_hash                        VARCHAR         DEFAULT NULL,
 
 -- metadata
     total_l2_tx_gas                   BIGINT          NOT NULL,
-    total_l2_tx_num                   BIGINT          NOT NULL,
-    total_l1_commit_calldata_size     BIGINT          NOT NULL,
+    total_l2_tx_num                   INTEGER         NOT NULL,
+    total_l1_commit_calldata_size     INTEGER         NOT NULL,
     total_l1_commit_gas               BIGINT          NOT NULL,
     created_at                        TIMESTAMP(0)    NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at                        TIMESTAMP(0)    NOT NULL DEFAULT CURRENT_TIMESTAMP,
